Upload IBDump to GitHub.
Description: Create ibdump.h file with the structs definitions.

Issue: 1898528
